orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "grpcio>=1.70",
    "grpcio-tools>=1.70",
    "grpcio-reflection>=1.70",
    "protobuf>=5.29",
]
